Changed version of Guest Additions to 5.0.10. Updated version.

+++ scripts/cleanup.sh
@@ -9,7 +9,9 @@ rm /var/lib/dhcp/*
 
 # Make sure Udev doesn't block our network
 echo "cleaning up udev rules"
-rm /etc/udev/rules.d/70-persistent-net.rules
+if [ -f /etc/udev/rules.d/70-persistent-net.rules ]; then
+	rm /etc/udev/rules.d/70-persistent-net.rules
+fi
 mkdir /etc/udev/rules.d/70-persistent-net.rules
 rm -rf /dev/.udev/
 rm /lib/udev/rules.d/75-persistent-net-generator.rules
